Making the multiselect selected option scrol int view when the selector is open.
It seems it is a long requested feature but not one was willing o
capable of opening a PR for it. That's an al to common situation I
suppose. In any case, rickus123 gave greate ideas about how to make a
workaround. I need to comment on that issue soon...

(This issue:
https://github.com/shentao/vue-multiselect/issues/616#issuecomment-1289265335)