Fixes the inedible amountPerUnit input!
Ha! I wrote unedible on purpose for this joke! Of course it was suppose
to uneditable!

In any case, I try a bunch of different approaches. I don't `if`s very
much. By chance I found an approach I think is really succint and
clear... I may be wrong about that though. It is in the eye of the
beholder I suppose...