Showing the selected option initially on MySelect.
There a few warning to take care of yet. And probably a good amount of
refactoring.

Also, there is a bug. When tab into the option list the first item is
focused but not "clicked"... The selected option remains null. Only
happens if you go to the first item and press enter.

I will probably look for a way to click the input (radio) on focus. This
may simplify other function handling keyboard navigation...