Watching for the "ready state" of a post operation only once.
So this specific code that updates the editableOperation and "redirects"
to the operation ID URL only "exists" when necessary.

I think it works but there is alot of refactoring to be done.
I may be able to extract some function from the view though I'm not sure
were to put the ones I'm able to extract or if I should extract
functions that return computed properties.

Also, add validation to the form.

Then improvements to the server side like ordering of operations...