I think I learned a LOT trying to make "react to search button without cache and to changes in the search params in the URL" work...
This was one my fiercest battle so far... This time for sure?

(reactivity and watches)

I will move the component to the directory with the same name soon.

It is very possible that there is alot to update and improve. I will
find out tomorrow (or later) when I get back to this code and see it
with a fresh pair of eyes.

For now, it works. (Or doens't! I will know when it is deployed and I
can take a look in my mobile.)