I moving the router reference I stored in the operations store in a new "common store". This solved a couple of type errors (this new store is a lot simpler).
And this way I should be able reuse it everywhere and not add one to
each store. That would be ridiculous.

This is all related to not being allowed to use composables outside
setup functions. Using stores for that looks like a loophole of sorts...

Anyway, et's see if it is going work for everything in the future. It is
definitely simpler and simpler is always better.