Simpler is always better! The crux of the problems I had before were because when landing on the /operations the URL search params was being changed...
And there was a watcher wathing the search params. Causing a deplication
of requests. I found a way to handle and make the watcher "wait" before
starting watching the search params. It was an interesting solution.
Clever even. But not necessary.

It is simpler and just as good or even better to not make that first
search params change to sync the URL with the datetime inputs.

This is the way.

I hope I did not break anything in the process! =P