Updating the PromiseState component to have a refresh "button" and display a count.
And updating the Operation and OperationList components to use this
improved PromiseState component and handle the refresh event.

I'm leavng alone a, likely prexisting, "bug": after entering the
operations page for the first time, when pressing the search button, it
just updates the search params and does make the request to get
operations.

From the second time on it works as intended. I may need to reconsider
how the search button works...