I'm starting to like this approach...
Nextg is implementing the "create" handler. I will have to consider
checking if the cached lists have the date of the created operation or
add the created operation to all lists anyway and regenerate the
byDate... This will kinda (softly) mess with the expectations set by the
set date range.

I should about it... Yeah.

I can have a function to check the new operation date is "in" the list
before regenerating byDate... Yeah... I will surely do that...